Refactor the service client interface.
This is deliberately similar to the interfaces for the other client
libraries (JS, Java). Obviously it is still namespaced under this
application; we are also still relying on GetServiceTokenAction as I
don't want to touch that code yet.

The only service interface for now is the ConfigDB. The others still
need refactoring into interface classes.

The Discovery interface exists, but is mostly a stub for now: it pulls
static URLs out of config specific to this application.