Search for the Edge host in the ConfigDB.
We need a record somewhere of which hosts are available and which
clusters they belong to. It makes more sense to me to put these in the
ConfigDB where they are accessible to other services and operators than
for them to be in the Manager's private database.

Edge hosts will need to be registered when they are added to the system.
This record includes a link to an Edge Cluster and a hostname.

If we are creating a Cell Gateway, search for an Edge Host matching the
hostname provided. It would be better to turn the hostname entry field
into a dropdown (or a per-building expanding menu, or something) with
hosts pulled from the list.