Update for better Kerberos integration and test environment setup
In this commit, the Kerberos integration was enhanced in the Dockerfile along with the addition of a script - get-keytab.sh to generate a keytab for application use. With the changes, the test environment can now be set up more accurately by duplicating the .env.example file to .env.testing. A line was added to .gitignore to ensure .env.testing is not tracked. The handling of Kerberos has been updated in the AuthenticateKerberosPrincipalAction.php file for better exception handling. Also, some of the irrelevant lines were removed from .env.example and README.md is updated to reflect these changes.