Refactor CreateNodeAction with improved condition handling
The CreateNodeAction has been refactored to handle conditions related to the app environment more effectively. The code was changed to better handle situations when the app is in 'local' environment or in 'testing' mode. Checks have been added to ensure that if the K8S_DEPLOY_WHEN_LOCAL environment variable is set to 'true', the local K8s config file is used, providing more flexibility and control over the configuration. This will make the application more robust by allowing for varied testing scenarios.