Updated CreateNodeAction.php to allow K8s deployment in local/testing environments.
Added an option to the condition checking the environment in the CreateNodeAction script. Previously, deployments to a Kubernetes cluster only happened if the app was not in a 'local' environment, unless specifically overrode. Now, deployments can also be avoided in the 'testing' environment. Also, the 'LOCAL_K8S_CONFIG_FILE' environment variable was replaced by 'LOCAL_KUBECONFIG', 'default'. This could be useful for testing configurations before deploying to production.