Add delete device functionality in DeviceEditor
We have added a new delete button in device editor's navigation. When clicked, it warns user about the destructive and irreversible action. On accepting, it takes a server call to delete the device and all its associated data and upon successful deletion, reroutes the user to homepage.

The changes are mainly contained within methods 'maybeDeleteDevice()' and 'deleteDevice ()' along with the addition of 'deleting' status in data object. This helps improve the manageability of devices for the user.