Update condition to handle null values in UpdateEdgeAgentConfigForNode
Changed the condition in UpdateEdgeAgentConfigurationForNodeAction to handle 'null' values. Previously, null values were being returned as is, but now they are interpreted as 'false'. This was done to ensure that false values for recordToDB are included in the config so that they can be picked up by the edge agents and interpreted for is_transient.