Implemented user permission management for node access
This commit introduces improvements and features that deal with node permission management. A user can now be granted or denied access to configure devices in a specific node, providing more security and control over the devices. Also, 'NodeUserController' and 'GrantUserPermissionToAccessNodeRequest' files were added to handle the functionality part while an overlay was introduced in the UI to manage permission from the front-end. A new http route was created to handle the new post request for granting user permission. The composer.json file was also updated with the 'ext-http' dependency. The change will give administrators finer control over individual node access permissions