Update DeviceEditor and SchemaGroup for dynamic object deletion
This update renames the `createDynamicSchemaObjects()` function to `updateDynamicSchemaObjects()` to better represent its function. It also includes a handler for `deleteObject` event in DeviceEditorOriginMapTab.vue and SchemaGroup.vue which executes `maybeDeleteObject()` for giving user confirmation on deletion. In addition, a `deleteObject()` function is created to perform actual deletion in model and schema after user confirmation. The 'canBeDeleted()` function is added in SchemaGroup.vue to decide if an object can be deleted. This update enhances user interaction with dynamic objects by allowing them to add and delete objects as needed.