Fix instantiated objects being created with new objects
The changes in this diff mainly consist of removing a console.log statement, adding a new function "removePropertiesWithPatternProperties", and calling that function on "schemaToInstantiate.properties".

The console.log was probably for debugging purposes, and has now been commented out.

The new function recursively checks for object properties that also have patternProperties and removes the properties where this is the case. This ensures that instantiated objects are not bloated with unnecessary new objects.

The new function is then called on "schemaToInstantiate.properties", making sure that the instantiated schema is nice and clean before being created.

In summary, these changes make sure that instantiated objects have all the unnecessary new objects removed and are created with a clean slate.