Update GitHub workflows and remove unused ones
The current workflows for building pull-requests have been updated to check for compatibility with the 'v1-stable' branch instead of the 'main' branch. This aligns better with current development practices.

The 'docker-publish.yml' and 'duster-fix-blame.yml' workflows were removed. 'docker-publish.yml' automated the use of Docker, but dependency changes rendered it inapplicable. 'duster-fix-blame.yml' was part of an experimental linting feature that is no longer being pursued.