Refactor auth action and Dockerfile, remove unnecessary logs, and add Ray configuration
An important change of this commit lies in the refactoring of execute method in GetServiceTokenAction. The new code is cleaner, more readable and more importantly, standard exceptions are thrown when necessary rather than relying on log or exit calls. This change makes it easier for error handling systems downstream to catch and process these exceptions.

Updates to Dockerfile are small yet crucial. These changes denote specifying Docker platform as linux/amd64 for build stages that can potentially offer better compatibility across different systems.

Unused logging and debugging lines have been removed from the system. This makes the code cleaner and reduces the risk of exposing potentially sensitive information in logs.

The inclusion of 'krb5.conf' in .gitignore helps ignore unnecessary files and keeps the repository clean. An update in docker-compose.yml makes it easy to build and run the application with Docker.

Additionally, a new ray.php file is added which is the configuration for Ray, a desktop app to debug Laravel applications.

Finally, there have been minor changes and corrections in JS and Vite configurations.