Add exception handling in device and node deletion
Added exception handling blocks to ensure any failures in particular steps of deleting a node or device doesn't stop the full process. Errors that occur during the update of EdgeAgentConfigurationForNode, deletion of KerberosKey, deletion of a deployment, or deletion of principal from AuthService and ACL entry are now caught and a relevant message is logged, instead of the system halting the process. This will allow for smoother execution and improve debuggability.